removing dead code from server.js, proper source code representations of emoji characters
